removed vcloudair,v57, getVdcOrg(vdc.go),and incorrect logging
The incorrect logging in vm.go and vapp.go were removed so that the user can run tests on the go-vcloud-director project with go test. I removed the v57  as Terraform does not use this at all. I removed all the vca instances including the function getVDCOrg which does not work on vcd. I also changed the struct K to be TestVCD since that is a more accurate name for it and removed struct S which is made for vca. Changed all references to struct S to be to TestVCD for clarity.
Signed-off-by: auppunda <auppunda@gmail.com>