Updates Issue#10 (no mocks for VM test)
In the mock version, this test used a predefined string to find the VM
and perform a simple comparison.

In this implementation, rather than forcing the user to provide yet
another parameter, we search for the given vApp (or the first available
one, if none was provided) and then we look for the first available VM.

Then, we perform the test as before, using the search by HREF and
comparing the newly found object fields.

Signed-off-by: Giuseppe Maxia