Initial Org Crud Implementation (#32)
* Initial Org Crud Implementation
First added the admin org type into types.go. This includes Vdc, Catalog, Network Lists, and settings. The settings variable is documented in types.go. Private functions added to aid in deletion of vdc, vapps, Catalogs, and Networks. Added Create in system.go, and another get function specifically for adminOrg. Added Update as well which updates it based on the adminorg variable passed in.

Signed-off-by: auppunda <auppunda@gmail.com>

* Removed one letter variables in functions added, changed test to check in tests
I removed one letter variables in the functions added for org creation. For the entire removal of these variables, it will be done with a separate pr. Also changed every test that used tests *C to be check *C

* Made org tests more thorough and made more precise variables names
Changed General to OrgGeneralSettings and VappTemplate to OrgVappTemplateSettings in types.go. Also made the create org test test more of the settings

Signed-off-by: auppunda <auppunda@gmail.com>

* Updated Testing.md
Added provider org for testing purposes

* Update issue #53 :Made an OrgOperations interface, and made adminOrg FindCatalog
Per Giuseppe's suggestion, I made an OrgOperations interface that contains all the functions similar to Orgs and adminOrgs so that we can use polymorphism for adminOrgs to be able to call Org functions. Also added to my adminOrg comment and removed the TODO comment. Also added a admin version of FindCatalog and added appropriate test to org_test.go

Signed-off-by: auppunda <auppunda@gmail.com>

* Fixed go vet issues

Signed-off-by: auppunda <auppunda@gmail.com>

* Changed one letter variable name to a better name
I changed b in create/update to xmlData, and u to orgCreateHREF

Signed-off-by: auppunda <auppunda@gmail.com>

* Added GetVdcByName to the OrgOperations interface
I added a adminorg version of GetVdcByName to org.go, added test as well

Signed-off-by: auppunda <auppunda@gmail.com>

* Added Documentation to FindCatalog function in org.go

Signed-off-by: auppunda <auppunda@gmail.com>

* Made variable names more expressive and documented test cases
I changed the loop variable names to represent what they actually are and added better documentation to the tests.

Signed-off-by: auppunda <auppunda@gmail.com>

* Updated issue #69 for org and system.go
Haven't introduced vcderror but have changed all get functions in org and system.go to return no error if it can't find the entity. I updated teh tests to try to look for an entity that doesn't exist.
Signed-off-by: auppunda <auppunda@gmail.com>