Implement Issue #77: dump failing HTTP operations
Implemented as a Fixed-size stack, based on standard Go double-linked
list. The stack is created with a size limit. When the limit is reached,
the oldest element is purged from the back end.

The stack is enabled when logging is disabled.
When there is a failure, the HTTP processing functions in api.go invoke
the dumping of the stack, which will briefly enable logging and write
down all elements in the stack (by default are 4: two requests and two
responses). The operations are listed in reverse order of execution
(most recent first) with a header that explains that we are emptying the
stack and which element number we are seeing.

Added documentation in the code and in LOGGING.md

Signed-off-by: Giuseppe Maxia