Initial Catalog CRUD operations (#57)
*Updated types.go with AdminCatalog definition

* Initial Catalog CRUD
Implemented Create, Update, Delete, and Get for Catalogs

* Added tests for Delete, Update, Get for catalogs
Added the tests for Delete, Update, and get

* Added Documentation to Catalog Delete/Update and AdminCatalog struct
Added Descriptive comments to AdminCatalog struct, and Delete/Update methods

* Added a more comprehensive definition of AdminCatalog
Per Giuseppe's concern, I added the missing elements to complete the AdminCatalog definition

* Added comments to the catalog tests, and cleaned up a variable name
* Fixed test error
* Made the hardcoded strings constants and added checks for if the entity exists
I added a check to see if the entity exists before creating it. If it does it deletes it and then goes on with the test. Also made all the names of the entities created during testing constants.

Signed-off-by: auppunda <auppunda@gmail.com>