Add option to skip logging of '/versions' API call
Since the '/versions' call can have a rather large response, which will
make the logs more difficult to search, a new variable is introduced,
'logging.SkipVersionsResponse`. When set, it will filter off the
response from the '/versions' call.

It is also possible to set the variable directly into the configuration
file for testing, or using an environment variable
(GOVCD_LOG_SKIP_VERSIONS)

Signed-off-by: Giuseppe Maxia <gmaxia@vmware.com>