Add workaround for disk update bug
The disk update has a problem when using a disk size that is less than
one MB. Even if the disk size is the same or bigger than the original
disk, when the size is less than one MB the update fails with an error
message saying that the update size is less than the original size.
To circumvent this error, I have introduced a check that requires at
least 1 MB for the disk size, both for creation and for updates.

Signed-off-by: Giuseppe Maxia <gmaxia@vmware.com>