Update copyright check with smarter algorithm
Instead of checking the file date, which could change because of
a fresh clone, we check either the file status (modified, not committed
yet) or the commit date using git. This way, we can get a meaningful
warning both in the developer's environment and in the automated check
by a CI tool.

Signed-off-by: Giuseppe Maxia <gmaxia@vmware.com>