Bug fixing 2018 12 (#135)
* Add option to skip logging of custom tags in API calls  

Since some tags in the API call response  can have a rather large text, which will
make the logs more difficult to search, a new function is introduced,
'logging.SkipApiTags`. When called, it will filter off a custom list of tags.

* Add option to only show the API calls from a custom list of functions in the logs.

* Add workaround for disk update bug

The disk update has a problem when using a disk size that is less than
one MB. Even if the disk size is the same or bigger than the original
disk, when the size is less than one MB the update fails with an error
message saying that the update size is less than the original size.
To circumvent this error, I have introduced a check that requires at
least 1 MB for the disk size, both for creation and for updates.

* Improve usability of `api.go`.

Change status code numbers to http constants to improve readability.

* Add more info to error message for unhandled error messages

* Fix regression in caller names for NewRequest

When NewRequest was refactored into two functions, we lost the ability
of showing the caller of a given API operation. To workaround this fact,
and to protect the logging from similar future updates, we log the call
stack instead of the latest function name.

* Improve logging details

Change caller function name into a list of caller functions from the stack
Add the ability of filtering off any tag from the response output
Add the ability of only logging from a custom list of functions

* Fix documentation and comments

* Improve copyright check to better match file year

The copyright check would fail if the year in the copyright line
is not the current one. Changed the check to accept the previous year
in the copyright, and only complain if the file was changed this year
but the copyright notice uses the previous year.

* Change variable names for log skip and inclusion

* Update copyright for recently committed files

* The detection for sysAdnmin user had a typo that inverted the results.

* Add explanation of how the copyright check works

Signed-off-by: Giuseppe Maxia <gmaxia@vmware.com>