Add persistence to cleanup list
The cleanup list is now contained in a file. When the test suite
terminates before it reaches the cleanup phase, the cleanup list
is preserved in a file that is unique for the vCD being used.
At the next test run, if that file is found, a clean up will be performed
before the tests starts.