Update OrgNetwork structure
The reason for this addition (VimPortGroupRef) is that when
we manipulate DHCP records, the API complains about this element
being missed. We don't need to add anything inside VimPortGroupRef,
but it needs to be there or we can't add DHCP definitions.

Note that, before this addition, DHCP was created automatically
using the leftwovers of static IP pool. For example, if the static
IP pool was 192.168.0.2-192.168.0.100, the DHCP pool, was
192.168.0.101-192.168.0.254.

Signed-off-by: Giuseppe Maxia <gmaxia@vmware.com>