Change refresh_token to use request body
Although the operations in go-vcloud-director are logged
with the sensitive information filtered off, when using
the API token in the URL, it gets recorded in the VCD
logs. To avoid this logging, which may result in
token disclosure when a support bundle is generated,
we change the refresh_token operation to use the request body
instead of URL parameters.

Signed-off-by: Giuseppe Maxia <gmaxia@vmware.com>