Adjust VM code (#509)
This code has a few tweaks for existing VM code.
* New function `vApp.AddRawVM` to support addition of VM from type `*types.ReComposeVAppParams` instead of functions such as `AddNewVMWithComputePolicy`, `AddNewVMWithStorageProfile`, `AddNewVM` which only accept some field values and need constant improvement.
* `vm.ChangeCPUAndCoreCount` which is almost the same as `vm.ChangeCPU` except that it accepts pointers instead of integers - this gives the option to specify only one of values the other being nil, which is more convenient in some workflows (including terraform-provider-vcd)
* Type `types.RecomposeVAppParamsForEmptyVm` had missing `PowerOn` field which is used to control VM power state on creation

Additional. Apparently we had a long existing inverse action for testing flag `GOVCD_IGNORE_CLEANUP_FILE`. It would perform cleanup for previous files only if `GOVCD_IGNORE_CLEANUP_FILE` is set (which is the reverse of its meaning). This PR patches condition check to match name of the flag.