Add org users to govcd test configuration
A new structure `[]tenants` is added to `TestConfig`, to allow
connecting as tenant user.
When the option `-vcd-test-org-user` (same as what is used in
terraform-provider-vcd) is se, the first tenant in the list is
selected for connection.
To use the second tenant (or an optional third one), we can
use the option `-vcd-connect-tenant=1` (or `-vcd-connect-tenant=2` if a
third one exists).

Signed-off-by: Giuseppe Maxia <gmaxia@vmware.com>